goodscenter 商品管理模块完成

+++ goodscenter-dal/src/main/java/com/mall/goodscenter/dal/dataobject/GoodsInfoDO.java
@@ -16,7 +16,7 @@ public class GoodsInfoDO extends BaseDO {
 
     private String goodsIntro;
 
-    private Long goodsCategoryId;
+    private Integer goodsCategoryId;
 
     private String goodsCoverImg;
